no new functionality, but refactoring for better arch.; moved RewardViewModel instantiation to where all other view models are instantiated.
