+++ app/src/main/java/com/kk/android/bayareanews/presentation/BayAreaNewsNavHost.kt
@@ -1,9 +1,11 @@
 package com.kk.android.bayareanews.presentation
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
@@ -13,6 +15,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import com.kk.android.bayareanews.R
 import com.kk.android.bayareanews.common.Constants
+import com.kk.android.bayareanews.data.TRHelper
 import com.kk.android.bayareanews.presentation.ui.common.Screen
 import com.kk.android.bayareanews.presentation.ui.common.WebViewScreen
 import com.kk.android.bayareanews.presentation.ui.home_screen.ContactInfoScreen
@@ -30,7 +33,6 @@ import kotlinx.coroutines.flow.MutableStateFlow
 
 @Composable
 fun BayAreaNewsNavHost(
-    rewardViewModel: RewardViewModel,
     navigationActions: BayAreaNewsNavigationActions,
     isExpandedScreen: Boolean,
     modifier: Modifier = Modifier,
@@ -40,12 +42,18 @@ fun BayAreaNewsNavHost(
     speechFlow: MutableStateFlow<String>?,
     onSpeechButtonClicked: () -> Unit
 ) {
-
+    val activityContext = LocalContext.current
+    val rewardViewModel = hiltViewModel<RewardViewModel>()
+    (rewardViewModel.trApi as TRHelper).context = activityContext
+    rewardViewModel.initRewards()
+    val showRewardScreenStateFlow = rewardViewModel.showRewardScreenStateFlow.collectAsState()
+    if (showRewardScreenStateFlow.value) {
+        navigationActions.navigateToRewards()
+    }
     NavHost(
         navController = navController,
         startDestination = startDestination
     ) {
-
         composable(
             Screen.HomeScreen.route
         ) {
